+++ core/src/main/kotlin/no/elg/hex/model/FastIslandMetadata.kt
@@ -3,6 +3,7 @@ package no.elg.hex.model
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.utils.Disposable
+import com.badlogic.gdx.utils.GdxRuntimeException
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonPropertyOrder
 import com.fasterxml.jackson.module.kotlin.readValue
@@ -63,7 +64,11 @@ class FastIslandMetadata(
   val preview: Texture?
     get() {
       if (internalPreviewTexture != null) return internalPreviewTexture
-      internalPreviewTexture = textureFromBytes(previewPixmap ?: return null)
+      try {
+        internalPreviewTexture = textureFromBytes(previewPixmap ?: return null)
+      } catch (e: GdxRuntimeException) {
+        Gdx.app.error("FastIslandMetadata", "Failed to load preview of $id!", e)
+      }
       return internalPreviewTexture
     }
 
@@ -76,11 +81,12 @@ class FastIslandMetadata(
   fun save() {
     require(id >= 0) { "Island id must be positive, is $id" }
     requireNotNull(previewPixmap) { "A preview have not been generated" }
-    requireNotNull(preview) { "Cannot save a preview that is not loadable, size of byte array is ${previewPixmap?.size}" }
     // ok to write to file when Hex.args.writeARtBImprovements is true as the app will exit afterwards
     val serializedThis = Hex.mapper.writeValueAsBytes(this)
     val compressed = compressXZ(serializedThis) ?: serializedThis
     if (Hex.mapEditor || Hex.args.writeARtBImprovements) {
+      // preview check only happens when we save it. For local saving we dont really care (user might have ran out of memory)
+      requireNotNull(preview) { "Cannot save a preview that is not loadable, size of byte array is ${previewPixmap?.size}" }
       require(modifier == PreviewModifier.NOTHING) { "Cannot save a modified preview in the map editor, is $modifier" }
       require(userRoundsToBeat == Island.NEVER_PLAYED) { "userRoundsToBeat must be ${Island.NEVER_PLAYED} when saving initial island, is $userRoundsToBeat" }