+++ CHANGES.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix ARtB not being reset when using edit tools
 * Fix capital placement considering the team of invisible hexagons
 * Fix quick loading not working more than once
+* Fix the game not being playable when low on memory
 
 ### Removed